Rewrite main gnuplot snipets to use macros
By moving the long 'plot \ ..' command to it's own
file, using the @col2 macro as a generic drop-in
for the y-axis, it is much simpler to append a
simulation directory (e.g. testing) for
visualization with the others.

This way, in the main plot file (e.g.,
restmass.cfg) we set col2="$4" and then
load 'plot-args.cfg', which is one substitution
instead of many. Maintaining the gnuplot snippets
is much less painful this way.