Remove un-needed settledisk-only files
It is much easier to adjust the paths config
file to just plot settle disk than it is to
maintain a dozen duplicate snippets for one
epoch only.

In any case, it only make sense to look at the
entropy and unbound electron fraction in the
settle disk epoch, so those two files have been
kept back.